refactor: split level 3 tests
